\name{concom-package}
\alias{concom-package}
\docType{package}
\title{
  Connected Components
}
\description{
  Provides a function which computes the connected components of an undirected 
  graph. 
}
\details{
  There are three functions in this package: \code{concom}, the main function, 
  which returns the connected components from the edges of the graph; 
  \code{concomFromMatAdj}, which returns the connected components from the 
  adjacency matrix of the graph; \code{concom3d}, which returns the connected 
  components of a triangle \strong{rgl} mesh.
}
\author{
Stéphane Laurent.

Maintainer: Stéphane Laurent <laurent_step@outlook.fr>
}
\keyword{ package }
