% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concom.R
\name{concomFromMatAdj}
\alias{concomFromMatAdj}
\title{Connected components from adjacency matrix}
\usage{
concomFromMatAdj(M)
}
\arguments{
\item{M}{adjacency matrix; it must be a square symmetric matrix with 
numeric or Boolean entries, whose non-zero or \code{TRUE} entries 
indicate the connections (connection between \code{i}-th vertex and 
\code{j}-th vertex if the entry is located at row \code{i} and 
column \code{j})}
}
\value{
The output is the same as the one of the \code{\link{concom}} 
  function.
}
\description{
Connected components of an undirected graph from its 
  adjacency matrix.
}
\examples{
matAdj <- rbind(
  c(0, 1, 0, 0),
  c(1, 0, 0, 0),
  c(0, 0, 0, 0),
  c(0, 0, 0, 1)
)
concomFromMatAdj(matAdj)
}
